add damping factory
